fix issue #305: Problem related to the visualization of the azimuth and elevation in the moutTui of Medicina is now solved. (#309)
The issue was due to the porting from 32 to 64 bits platform, particularly the length of long integers (8 bytes).
During the decoding of the status message from the ACU the elevation and azimuth values were bitwise copied from the buffer
into a long variable. The elevaion and azimuth i the ACU status is 4 bytes long so the most significant part of the internal
variable wo not initialized.