Fix issues from 366 to 369. See changelog below. (#371)
* Fix issues from 366 to 369. See changelog below.

Fix #366, now the PyCalmux component is kept alive after it's initially instanced. This prevents setting the default configuration to the controlled device every time the component is referenced after it was previously released.
Fix #367, PyCalmux connection to its controlled device is now handled in a smarter way to prevent failing to open a new connection right after a previous one was closed.
Resolve #368, the PyCalmux setup method now can be called from the OperatorInput using the `calmux=<backend_name>` command.
Resolve #369, the Noto StationProcedures now execute the `calmux=TotalPower` command. It means that when a procedure command is issued (i.e.: `setupKKC`), the CalMux device sets the TotalPower backend as the source of the calibration noise mark by default.

* Fixed some indentation consistency in CHANGELOG.md

* Added PyCalmuxContainer to Noto discosStartup.xml