Fix #377, TimeTaggedCircularArray::selectPoint refactored
Now the method waits for a maximum of 2 seconds for a new point to be inserted into the array when a set of coordinates with a too recent timestamp is requested. If no point is received, it return the last set of known coordinates and a boolean value of `false` indicating that the returned pair of coordinates is uncertain. This boolean value can be useful in case we want to log the event from the caller module (i.e. from the `AntennaBoss` module).