Splitted WeatherStation::getData in two different methods. See below.
The two methods are `getData`, which returns the last read weather parameters immediately, and `updateData`, which performs the request to the devIOs tu update every weather parameter.
The `updateData` method is now called periodically (every 2 seconds) from the `UpdatingThread` (the former `WindCheckerThread`). This thread also checks if the wind is too strong and automatically stows the antenna.