Fix #431, dropped the use of the SecureArea in the AntennaBoss WatchingThread (#432)
* Fix #431, dropped the use of the SecureArea in the AntennaBoss WatchingThread

The AntennaBoss WatchingThread is now free to run without being blocked by a SecureArea. This allows both the WatchingThread and the WorkingThread to be executed in parallel. This was tested using the SRT ACU simulator. The test has been running for more than one hour without issues. Further testing with the SRT will be performed soon (2019/07/25). This SHOULD NOT break the MedicinaMount, but it was not tested, so I'm not 100% sure.

P.S.: I also tried a setup with both threads not protected by a SecureArea, but after a while the tracking behaved unexpectedly, sending to the ACU a set of coordinates with an elevation of more than 2000 degrees. This behavior was most likely caused by the WorkingThread not protected by the SecureArea.

* Fix #431, improved synchronization between AntennaBoss threads

The load/unloadMount, load/unloadRefraction load/unloadPointingModel methods are now thread safe in order to minimize (unfortunately not avoid completely) concurrency errors between the AntennaBoss WatchingThread and WorkingThread
The TimeTaggedCircularArray object is now thread safe.