Fix #437, slightly reviewed USDs positions updating strategy
Now the elevation is not checked whenever a new position has to be commanded. The check is made for every USD. If the position to be commanded is different than the current one it gets commanded, ignored otherwise. This prevents sending the same position over and over again to the USDs, avoiding to busy up the LAN (and the receiving USD) with commands that will have no effects whatsoever.

Fix #438, removed the SecureArea locking mechanism from every method that end up calling `oneway void` methods of the USDs. Since the `oneway void` methods are asynchronous, and since most of the times 1116 methods of this kind are called together, their execution probably completes after the SecureArea is released, making it pointless to lock it in the first place. The level in which a SecureArea provides actual synchronization is in the LAN component (since every LAN is shared among multiple USDs).

Is mandatory to test this branch thoroughly with the hardware before commiting it to production line. It has been tested over the weekend with the simulators without issues, but this is not enough.