Fix #437, filtered the new commanded positions to the SRT AS USDs. (#449)
* Fix #437, filtered the new commanded positions to the SRT AS USDs.

A check on the last commanded step is performed before commanding a new step to each SRT active surface's USD. This will allow to avoid sending the same position multiple times to the USDs, introducing less stress on the LANs and permitting the outermost circles' USDs to be updated in a faster way. This also means that the GUI will have to wait a lot less for a response from the AS Boss when asking the status of the USDs.