Issue #447, the SRT AS GUI now takes into account the elapsed time for each iteration
It means that after the iteration is completed it goes to sleep for an amount of milliseconds equal to 100ms minus the elapsed time since the start of the current iteration. This will result in a sleep time of 0 whenever the elapsed time for the requests to the AS boss is greater or equal to 100ms. The GUI is still lagging though, so this does not fix the issue completely.