Fix #487, moved the start time acquisition after the SecureArea lock
When the SecureArea is not locked by any other thread, locking it requires virtually no time (a few microseconds). So now the locking elapsed time is not taken into account anymore. In this way the total execution time for a single iteration is equal to Tlock + Texecution, with Tlock usually negligible, or greater than a second whenever a new profile is chosen (due to a Wait(1000000) in the other locking thread).