Fix #447, the SRT AS GUI has been reworked in order to account for some latency
Now the SRT Active Surface GUI Client asks for the complete status of the active surface, instead of asking a single USD status at a time. This results on fewer calls to the AS Boss component, leaving more CPU time for the Boss' Working Thread (when in tracking mode). This allows us to consider shortening the working thread period (now it's 0.5 seconds) in order to achieve higher precision.
Also, the code of the GUI Client has been significally simplified by reusing the same code for all the actuator buttons (previously each one had its own methods).