Fix #614, fix #615. See changelog below. (#616)
* Fix #614, fix #615. See changelog below.

Issue #614: added the simulation .xml file for the ACS command center. It is automatically used when the testing CDB is in use.
Issue #615: changed output directory for Sardara NODATA files when the testing CDB is in use.

* Fix #614, now the correct environment is always loaded even when started graphically

This fixes a wrong behavior that showed up when launching discos from a .desktop file
The .desktop file loaded the environment variables at graphical session startup, so any environment update made by the discos-set command was simply ignored.