Fix #782, improved SRT MS singleton socket and tests
The socket is now set to non-blocking, therefore some checks were added in order to accomodate to this change.
Fixed a wrong behavior regarding the singleton design pattern. Previously it was possible to call the getInstance(ip_address, port) method multiple times with multiple addresses, but only the first instance was returned, meaning that only the first instance was actually opened to the correct IP address and port. Now if the user tries to open a second socket on a different IP address and port an exception is raised.