First working implementation with the actual hardware
Libraries and components were slightly reworked and tweaked in order to work correctly with the current hardware status.
The gregorian cover movement was took out of the setup and park procedures, since it is currently locked in place and won't move. We will add it again later.
All the currently unavailable minor servos components were took out from the Boss control by commenting them. They can still be accessed via objexp and check their status.
The GFR was took out as well since it keeps blocking during movements. For now it has to be moved manually, it can be done via objexp as well, but resetting the errors from VBrain.
I rewrote the SRPProgramTrackTest.cpp in order to test the behavior of the SRP with new libraries. The test acts just like before.
Other tests need to be reworked, keeping the same logic behind.
Further investigation must be put into the scanning procedure, in particular when the recorder retrieves the coordinates.