SRT new Minor Servos (#852)
* Fix #672, fix #637, some improvements for the SRTMinorServoCommandLibrary (#674)

Fix #672, wrote all the commands in capital letters
Fix #673, added the OFFSET command to the library
Also, fixed changed the start_time parameter to be written as an integer

* Fix #646, wrote a Python wrapper for the `SRTMinorServoCommandLibrary` (#650)

* Fix #646, wrote a Python wrapper for the `SRTMinorServoCommandLibrary`

* Fix #646, Added a python module that uses the wrapper library

It has been commented in order to replicate the behavior in any future python wrapped C++ library

* Fix #646, fixed a few small things regarding some include and the Makefile

* Fix #646, added documentation about the Python wrapper

Also, fixed some small inconsistencies with the new C++ implementation

* Issue #637, removed an unneeded include statement

* Fix #684, fixed tests for the SRTMinorServoCommandLibrary and its python wrapper (#685)

* Fix #771, first implementation of parseAnswer and related tests (#772)

* Issue #771, first implementation of parseAnswer and related tests

Right now the function is not robust, it should check for errors in the answer format

* Fix #771, complete implementation of `parseAnswer` function

Tests were updated as well

* Moved old SRT Minor Servos to another directory

* Fix #782, wrote a singleton socket class for SRT new minor servos (#783)

* Fix #782, wrote a singleton socket class for SRT new minor servos

This is the first implementation, connection checks for the socket should be added in order to avoid some disconnection errors

* Fix #782, improved SRT MS singleton socket and tests

The socket is now set to non-blocking, therefore some checks were added in order to accomodate to this change.
Fixed a wrong behavior regarding the singleton design pattern. Previously it was possible to call the getInstance(ip_address, port) method multiple times with multiple addresses, but only the first instance was returned, meaning that only the first instance was actually opened to the correct IP address and port. Now if the user tries to open a second socket on a different IP address and port an exception is raised.

* Updated the SRTMinorServoCommandLibrary

* Added a test for SRP Program Track

* Some minor fixes for the SRTMinorServoCommandLibrary

* Updated SRPProgramTrackTest

* Improved SRP tests

* Updated SRPProgramTrackTest

* Updated SRPProgramTrackTest

Added a script to plot trajectories

* Added .gitignore in order to ignore test folders

* Updated SRPProgramTrackTest

Added a separate trajectories test

* Updated SRPProgramTrackTest.cpp

* Fix SRTMinorServoCommandLibrary programTrack start_time issue

* Minor fixes for SRTMinorServoCommandLibrary

* Updated SRP and Derotator programTrack tests

* Small update for plotting scripts

* Updated SRPProgramTrackTest.cpp

* First skeleton of combined SRP and Derotator tests

* Added sine wave tests

* Added combined test and plotting script

* Included rotations in sine wave tests

* Updated new minor servos tests

* Minor Servos update

* Some updates

* Updated CDB and schemas

* Updated SRTMinorServoLibraries

* Updated SRTMinorServoSocket and tests

* Commented new interfaces

* Updated CDB schemas and configurations

* Updated SRTMinorServo interfaces

* Updated SRT MinorServo Libraries and tests

* Commented new SRT minor servo files.

Commented pretty much every file. I will perform some more tests with the simulators before tweaking the system to be used with the current Leonardo minor servo system status.

* Fixed a couple bugs in SRTMinorServoBoss component

* Updated procedures in order to use new minor servos

* Fixes STOW command for the gregorian cover

* Minor improvement on the MS motion status DevIO

* Fix #834, avoided Leonardo Minor Servo SETUP command

* Updated CDBs, getting ready for production

* Updates for SRTMinorServos

* Updated SRTMinorServo setup and park procedures

* Updated SRTMinorServoLibrary

* Updates for the SRTMinorServos

* First working implementation with the actual hardware

Libraries and components were slightly reworked and tweaked in order to work correctly with the current hardware status.
The gregorian cover movement was took out of the setup and park procedures, since it is currently locked in place and won't move. We will add it again later.
All the currently unavailable minor servos components were took out from the Boss control by commenting them. They can still be accessed via objexp and check their status.
The GFR was took out as well since it keeps blocking during movements. For now it has to be moved manually, it can be done via objexp as well, but resetting the errors from VBrain.
I rewrote the SRPProgramTrackTest.cpp in order to test the behavior of the SRP with new libraries. The test acts just like before.
Other tests need to be reworked, keeping the same logic behind.
Further investigation must be put into the scanning procedure, in particular when the recorder retrieves the coordinates.

* Updated logging for SRTMinorServos

* Updated SRTMinorServo socket disconnect logging

* Fixed small bug in servos axes naming, improved logging

* Removed most of get_sync calls

* Fixed a couple more things

* Refactored SRTMinorServos

* Updated SRTMinorServoAnswerMap socket handling

* Added some log for the SRTMinorServoAnswerMap

* Updated branch

* Updated AntennaBoss component retrieving method

* Updated CCG procedure

* Fix #839, set a longer timeout for SRT minor servos setup and park

* Updated CHANGELOG.md