Updated telescopePark procedure commands order
With this order, the gregorian cover is started parking when the antenna is already at 89.8 degrees of elevation.
This will avoid getting an error on a gregorian cover axis pre limit.
Unfortunately, this will not fix a change in configuration from primary to gregorian or BWG and vice versa.
That will need to be addressed by opening a ticket.