Fix #906, various enhancements for Schedule parsing
This additions modify the IRATools::getNextToken function in the following way:
- a std::vector<char> can be passed as delimiter (function override). The function checks if any of them is present in the string and behaves according to the old implementation;
- the behavior has been enhanced so that a boolean (which defaults to false) tells the function to keep cycling through the string if multiple occurrencies of one of the delimiters is found one after the other. This allows the function to behave in the same manner whether a string contains one or more '\t' characters or a series of whitespaces ' '.
The Schedule class has been modified according to these additions so that it can now parse a schedule with multiple tabs as well as whitespaces as token separators. A test has been introduced in such a way that checks if the 2 schedule.tmpl and schedule_ws.tmpl files produce the same output when checked with the `testSched` executable.
ScheduleChecker now also checks whether the caller user is the same as the PROJECT keyword. This alerts the observers that copied a schedule from a previous project to be aware that DISCOS will reject the given schedule when trying to launch it, saving time.