Fix #908, set flag to speed up AS GUIs build (#930)
Every time we try to build a AS GUI, the process tries to build the objects with the flag -fvar-tracking-assignments enabled by default. The build process with this flag is never successful and the process prints a warning similar to:
"variable tracking size limit exceeded with -fvar-tracking-assignments, retrying without"
By disabling the flag in the makefile, we skip the first, always unsuccessful, attempt, saving a lot of time.